Switch LingBot-VLA 2.0 inference to fp16 and report the full metric set

+++ examples/offline_inference/lingbot_vla_v2/lingbot_vla_v2.py
@@ -42,7 +42,8 @@ def parse_args() -> argparse.Namespace:
     parser.add_argument("--model", required=True, help="Prepared model directory")
     parser.add_argument("--prompt", default="pick up the object")
     parser.add_argument("--seed", type=int, default=0)
-    parser.add_argument("--dtype", default="bfloat16")
+    # fp16 over bf16: see `spikes/lingbot_vla_v2/PHASE7_NUMERICS.md`.
+    parser.add_argument("--dtype", default="float16")
     return parser.parse_args()
